service--llm-hub/pyproject.toml
[project]
name = "llm-hub"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2",
    "pydantic-settings>=2",
    "loguru>=0.7",
    "python-dotenv>=1",
    "tomli>=2",
    "fastapi>=0.100",
    "httpx>=0.25",
    "uvicorn>=0.30",
    "openai>=2",
    "anthropic>=0.21.0",
    "gpt2giga>=0.1",
    "dishka~=1.8.0",
    "mmar-mapi~=1.6.0",
    "mmar-mimpl~=1.1.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
    "fire~=0.7.1"
]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
pythonpath = [ "src" ]

[pytest]
python_files = [ "test_*.py" ]
testpaths = [ "tests" ]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "C90",    # mccabe (complexity)
    "RUF",    # ruff-specific rules
    "TID",    # flake8-tidy-imports
    "ICN",    # flake8-import-conventions
    "PTH",    # flake8-use-pathlib
    "SIM",    # flake8-simplify
    "S",      # flake8-bandit (security)
    "A",      # flake8-builtins
    "PLC",    # pylint
    "PLE",    # pylint
    "PLR",    # pylint
    "PLW",    # pylint
    "TRY",    # tryceratops
]
ignore = [
    "RUF001", # ambiguous-unicode-character-string
    "S101",   # assert-used
    "S102",   # exec-used
    "S104",   # bind-all-interfaces
    "S108",   # hardcoded-temp-file
    "S113",   # request-without-timeout
    "S301",   # pickle
    "PLR0913", # too-many-arguments
    "PLR0915", # too-many-statements
    "PLR0912", # too-many-branches
    "PLR2004", # magic-value-comparison
    "TRY003", # long-exception-messages
    "TRY300", # raise-within-try
    "TRY301", # raise-within-try
    "TRY400", # error-instead-of-exception
    "N814",   # camelcase-imported-as-lowercase
]

[tool.ruff.lint.mccabe]
max-complexity = 20

[tool.ruff.lint.isort]
known-first-party = ["llm_hub"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.mypy]
explicit_package_bases = true
mypy_path = "src"
plugins = ["pydantic.mypy"]
